Fix Error getting Wi-Fi password
iwgetid(8) was used to find out the ESSID of the wireless network that is
currently used. Unfortunately iwgetid needs root / sudo rights, that's
why I replaced the code with nmcli.

nmcli uses connection IDs / name to identify the connection. One can not
assume that the connection name is the same as the wifi name so I
introduced a new variable `connection_name` to be used in nmcli to find
out the wifi password.

I am not sure how a connection without wifi-password looks like or
rather what nmcli retruns in that case. We need to test that case.

Closes: thanosgn/wifi-share#4