Fix Error getting Wi-Fi connections
Displaying the list of stored Wi-Fi networks (SSIDs) to choose from
failed. The problem on this side is that the type of the Wi-Fi network
connections is '802-11-wireless' which does not contain 'wifi'.

To be backward compatible we search for "wireless" and "wifi".

Then the second problem is that it was assumed that the network
connection does not contain white spaces. To overcome that limitation
we changed to the "terse" output format which makes it easier to grep
for the connection name (the fields are delimited with ":").

And finally "nmcli connection show" does not show SSIDs but "connection
names", so we have to retrieve the wifi_name from the connection_name.

Closes: thanosgn/wifi-share#7